refactor(authentication): remove routing manager from oauth constructor add(authentication): declare route function to all handler classes refactor(authentication): cleanup routing declaration
