+++ modules/authentication/src/handlers/phone.ts
@@ -17,16 +17,17 @@ import { TokenType } from '../constants/TokenType';
 import { ISignTokenOptions } from '../interfaces/ISignTokenOptions';
 import moment = require('moment');
 import { Cookie } from '../interfaces/Cookie';
+import { IAuthenticationStrategy } from '../interfaces/AuthenticationStrategy';
 
-export class PhoneHandlers {
+export class PhoneHandlers implements IAuthenticationStrategy{
   private sms: SMS;
   private initialized: boolean = false;
 
-  constructor(private readonly grpcSdk: ConduitGrpcSdk, private readonly routingManager: RoutingManager) {
+  constructor(private readonly grpcSdk: ConduitGrpcSdk) {
 
   }
 
-  async validate(): Promise<Boolean> {
+  async validate(): Promise<boolean> {
     const config = ConfigController.getInstance().config;
     let errorMessage = null;
     await this.grpcSdk.config
@@ -44,8 +45,8 @@ export class PhoneHandlers {
     }
   }
 
-  async declareRoutes() {
-    this.routingManager.route(
+  async declareRoutes(routingManager: RoutingManager) {
+    routingManager.route(
       {
         path: '/phone',
         action: ConduitRouteActions.POST,
@@ -60,7 +61,7 @@ export class PhoneHandlers {
       }),
       this.authenticate.bind(this),
     );
-    this.routingManager.route(
+    routingManager.route(
       {
         path: '/phone/verify',
         action: ConduitRouteActions.POST,
