+++ modules/authentication/src/interfaces/AuthenticationStrategy.ts
@@ -1,0 +1,6 @@
+import { RoutingManager } from '@conduitplatform/grpc-sdk';
+
+export interface IAuthenticationStrategy {
+  validate(): Promise<boolean>;
+  declareRoutes(routingManager: RoutingManager, config: any): void;
+}
