+++ modules/authentication/src/routes/routes.ts
@@ -69,257 +69,81 @@ export class AuthenticationRoutes {
       .catch((e) => (errorMessage = e));
 
     if (phoneActive && !errorMessage) {
-      await this.phoneHandlers.declareRoutes();
+      await this.phoneHandlers.declareRoutes(this._routingManager);
     }
 
     let authActive = await this.localHandlers
       .validate()
       .catch((e) => (errorMessage = e));
     if (!errorMessage && authActive) {
-      const fields = User.getInstance().fields;
-      delete fields.hashedPassword;
-      this._routingManager.route(
-        {
-          path: '/local/new',
-          action: ConduitRouteActions.POST,
-          description: 'Creates a new user using email/password.',
-          bodyParams: {
-            email: ConduitString.Required,
-            password: ConduitString.Required,
-          },
-          middlewares: [],
-        },
-        new ConduitRouteReturnDefinition('RegisterResponse', fields),
-        this.localHandlers.register.bind(this.localHandlers));
-
-      this._routingManager.route(
-        {
-          path: '/local',
-          action: ConduitRouteActions.POST,
-          description: `Login endpoint that can be used to authenticate. 
-              If 2FA is used for the user then instead of tokens 
-              you will receive a message indicating the need for a token from the 2FA mechanism.`,
-          bodyParams: {
-            email: ConduitString.Required,
-            password: ConduitString.Required,
-          },
-        },
-        new ConduitRouteReturnDefinition('LoginResponse', {
-          userId: ConduitString.Optional,
-          accessToken: ConduitString.Optional,
-          message: ConduitString.Optional,
-          refreshToken: ConduitString.Optional,
-        }),
-        this.localHandlers.authenticate.bind(this.localHandlers),
-      );
-
-      let emailOnline = false;
-      await this.grpcSdk.config.moduleExists('email')
-        .then(_ => {
-          emailOnline = true;
-        })
-        .catch(_ => {
-        });
-      if (emailOnline) {
-        this._routingManager.route(
-          {
-            path: '/forgot-password',
-            action: ConduitRouteActions.POST,
-            bodyParams: {
-              email: ConduitString.Required,
-            },
-          },
-          new ConduitRouteReturnDefinition('ForgotPasswordResponse', 'String'),
-          this.localHandlers.forgotPassword.bind(this.localHandlers),
-        );
-
-        this._routingManager.route(
-          {
-            path: '/reset-password',
-            action: ConduitRouteActions.POST,
-            description: `Used after the user clicks on the 'forgot password' link and
-                 requires the token from the url and the new password`,
-            bodyParams: {
-              passwordResetToken: ConduitString.Required,
-              password: ConduitString.Required,
-            },
-          },
-          new ConduitRouteReturnDefinition('ResetPasswordResponse', 'String'),
-          this.localHandlers.resetPassword.bind(this.localHandlers),
-        );
-      }
-
-      this._routingManager.route(
-        {
-          path: '/local/change-password',
-          action: ConduitRouteActions.POST,
-          description: `Changes the user's password but requires the old password first.
-                 If 2FA is enabled then a message will be returned asking for token input.`,
-          bodyParams: {
-            oldPassword: ConduitString.Required,
-            newPassword: ConduitString.Required,
-          },
-          middlewares: ['authMiddleware'],
-        },
-        new ConduitRouteReturnDefinition('ChangePasswordResponse', 'String'),
-        this.localHandlers.changePassword.bind(this.localHandlers),
-      );
-
-      this._routingManager.route(
-        {
-          path: '/hook/verify-email/:verificationToken',
-          action: ConduitRouteActions.GET,
-          description: `A webhook used to verify user email. This bypasses the need for clientid/secret`,
-          urlParams: {
-            verificationToken: ConduitString.Required,
-          },
-        },
-        new ConduitRouteReturnDefinition('VerifyEmailResponse', 'String'),
-        this.localHandlers.verifyEmail.bind(this.localHandlers),
-      );
-
-      if (config.twofa.enabled) {
-        this._routingManager.route(
-          {
-            path: '/local/twofa',
-            action: ConduitRouteActions.POST,
-            description: `Verifies the 2FA token.`,
-            bodyParams: {
-              email: ConduitString.Required,
-              code: ConduitString.Required,
-            },
-          },
-          new ConduitRouteReturnDefinition('VerifyTwoFaResponse', {
-            userId: ConduitString.Optional,
-            accessToken: ConduitString.Optional,
-            refreshToken: config.generateRefreshToken ? ConduitString.Required : ConduitString.Optional,
-            message: ConduitString.Optional,
-          }),
-          this.localHandlers.verify2FA.bind(this.localHandlers),
-        );
-
-        this._routingManager.route(
-          {
-            path: '/local/enable-twofa',
-            action: ConduitRouteActions.UPDATE,
-            description: `Enables a phone based 2FA method for a user and 
-                requires their phone number.`,
-            middlewares: ['authMiddleware'],
-            bodyParams: {
-              phoneNumber: ConduitString.Required,
-            },
-          },
-          new ConduitRouteReturnDefinition('EnableTwoFaResponse', 'String'),
-          this.localHandlers.enableTwoFa.bind(this.localHandlers),
-        );
-
-        this._routingManager.route(
-          {
-            path: '/local/verifyPhoneNumber',
-            action: ConduitRouteActions.POST,
-            description: `Verifies the phone number provided for the 2FA mechanism.`,
-            middlewares: ['authMiddleware'],
-            bodyParams: {
-              code: ConduitString.Required,
-            },
-          },
-          new ConduitRouteReturnDefinition('VerifyPhoneNumberResponse', 'String'),
-          this.localHandlers.verifyPhoneNumber.bind(this.localHandlers),
-        );
-
-        this._routingManager.route(
-          {
-            path: '/local/disable-twofa',
-            action: ConduitRouteActions.UPDATE,
-            description: `Disables the user's 2FA mechanism.`,
-            middlewares: ['authMiddleware'],
-          },
-          new ConduitRouteReturnDefinition('DisableTwoFaResponse', 'String'),
-          this.localHandlers.disableTwoFa.bind(this.localHandlers),
-        );
-
-        this._routingManager.route(
-          {
-            path: '/local/change-password/verify',
-            action: ConduitRouteActions.POST,
-            description: `Used to provide the 2FA token for password change.`,
-            bodyParams: {
-              code: ConduitString.Required,
-            },
-            middlewares: ['authMiddleware'],
-          },
-          new ConduitRouteReturnDefinition('VerifyChangePasswordResponse', 'String'),
-          this.localHandlers.verifyChangePassword.bind(this.localHandlers),
-        );
-      }
+      await this.localHandlers.declareRoutes(this._routingManager, config);
       enabled = true;
     }
     errorMessage = null;
 
-
     serverConfig = await this.grpcSdk.config.getServerConfig();
-    this.facebookHandlers = new FacebookHandlers(this.grpcSdk, this._routingManager,
+    this.facebookHandlers = new FacebookHandlers(this.grpcSdk,
       new FacebookSettings(this.grpcSdk, config, serverConfig.url));
     authActive = await this.facebookHandlers
       .validate()
       .catch((e) => (errorMessage = e));
     if (!errorMessage && authActive) {
-      this.facebookHandlers.declareRoutes();
+      this.facebookHandlers.declareRoutes(this._routingManager);
       enabled = true;
     }
 
-    this.googleHandlers = new GoogleHandlers(this.grpcSdk, this._routingManager,
+    this.googleHandlers = new GoogleHandlers(this.grpcSdk,
       new GoogleSettings(this.grpcSdk, config, serverConfig.url));
     errorMessage = null;
     authActive = await this.googleHandlers
       .validate()
       .catch((e) => (errorMessage = e));
     if (!errorMessage && authActive) {
-      this.googleHandlers.declareRoutes();
+      this.googleHandlers.declareRoutes(this._routingManager);
       enabled = true;
     }
 
-    this.githubHandlers = new GithubHandlers(this.grpcSdk, this._routingManager,
+    this.githubHandlers = new GithubHandlers(this.grpcSdk,
       new GithubSettings(this.grpcSdk, config, serverConfig.url));
     errorMessage = null;
     authActive = await this.githubHandlers
       .validate()
       .catch((e) => (errorMessage = e));
     if (!errorMessage && authActive) {
-      this.githubHandlers.declareRoutes();
+      this.githubHandlers.declareRoutes(this._routingManager);
       enabled = true;
     }
 
-    this.slackHandlers = new SlackHandlers(this.grpcSdk, this._routingManager,
+    this.slackHandlers = new SlackHandlers(this.grpcSdk,
       new SlackSettings(this.grpcSdk, config, serverConfig.url));
     errorMessage = null;
     authActive = await this.slackHandlers
       .validate()
       .catch((e) => (errorMessage = e));
     if (!errorMessage && authActive) {
-      this.slackHandlers.declareRoutes();
+      this.slackHandlers.declareRoutes(this._routingManager);
       enabled = true;
     }
 
-    this.figmaHandlers = new FigmaHandlers(this.grpcSdk, this._routingManager,
+    this.figmaHandlers = new FigmaHandlers(this.grpcSdk,
       new FigmaSettings(this.grpcSdk, config, serverConfig.url));
     errorMessage = null;
     authActive = await this.figmaHandlers
       .validate()
       .catch((e) => (errorMessage = e));
     if (!errorMessage && authActive) {
-      this.figmaHandlers.declareRoutes();
+      this.figmaHandlers.declareRoutes(this._routingManager);
       enabled = true;
     }
 
-    this.microsoftHandlers = new MicrosoftHandlers(this.grpcSdk, this._routingManager,
+    this.microsoftHandlers = new MicrosoftHandlers(this.grpcSdk,
       new MicrosoftSettings(this.grpcSdk, config, serverConfig.url));
     errorMessage = null;
     authActive = await this.microsoftHandlers
       .validate()
       .catch((e) => (errorMessage = e));
     if (!errorMessage && authActive) {
-      this.microsoftHandlers.declareRoutes();
+      this.microsoftHandlers.declareRoutes(this._routingManager);
       enabled = true;
     }
 
@@ -348,14 +172,14 @@ export class AuthenticationRoutes {
 
       enabled = true;
     }
-    this.twitchHandlers = new TwitchHandlers(this.grpcSdk, this._routingManager,
+    this.twitchHandlers = new TwitchHandlers(this.grpcSdk,
       new TwitchSettings(this.grpcSdk, config, serverConfig.url));
     errorMessage = null;
     authActive = await this.twitchHandlers
       .validate()
       .catch((e) => (errorMessage = e));
     if (!errorMessage && authActive) {
-      this.twitchHandlers.declareRoutes();
+      this.twitchHandlers.declareRoutes(this._routingManager);
       enabled = true;
     }
     if (enabled) {
